Stabilize ROI identity, geometry, and tooling
Fixes multiple Track A correctness gaps across ROI handling: unknown layer identities are now minted per layer/source-safe, affine/world mapping uses napari transforms, unsupported ROI geometries raise explicit errors, and ROI mask payload decoding is hardened (size/run/zlib validation with measured codec sizing). It also centralizes processor provenance attachment with preserves_grid semantics, ensures display-layer components inherit parent identity, routes ROI manager mutations through command logging with UID-based row identity, hardens viewer tool ownership/callback release, and reduces per-ROI float64 conversions for colocalization/PSF/stat paths.

+++ imswitch/improcess/analysis/colocalization.py
@@ -141,8 +141,11 @@ def _extract_values(
     image_b: np.ndarray,
     roi: ROIRecord | tuple[int, int, int, int] | None,
 ) -> tuple[np.ndarray, np.ndarray, tuple[int, int, int, int] | None]:
-    a = np.asarray(image_a, dtype=np.float64)
-    b = np.asarray(image_b, dtype=np.float64)
+    # Not converted to float64 up front: this runs once per ROI, and casting
+    # two full images each time is the cost the local rasteriser removes.
+    # Slices are cast instead, at the point of use.
+    a = np.asarray(image_a)
+    b = np.asarray(image_b)
     if a.ndim != 2 or b.ndim != 2:
         raise ValueError(f"Colocalization expects two 2D images, got {a.shape} and {b.shape}")
     if a.shape != b.shape:
@@ -155,18 +158,18 @@ def _extract_values(
         # segmentation mask fell through to the bounding rectangle and
         # correlated the box instead of the region.
         local, slices = roi_mask_local(roi, a.shape)
-        av = a[slices][local]
-        bv = b[slices][local]
+        av = a[slices][local].astype(np.float64, copy=False)
+        bv = b[slices][local].astype(np.float64, copy=False)
         bounds = roi_bounds(roi)
     else:
         bounds = roi.bounds if isinstance(roi, ROIRecord) else roi
         if bounds is None:
-            av = a.ravel()
-            bv = b.ravel()
+            av = a.ravel().astype(np.float64, copy=False)
+            bv = b.ravel().astype(np.float64, copy=False)
         else:
             r0, r1, c0, c1 = _clip_bounds(bounds, a.shape)
-            av = a[r0:r1, c0:c1].ravel()
-            bv = b[r0:r1, c0:c1].ravel()
+            av = a[r0:r1, c0:c1].ravel().astype(np.float64, copy=False)
+            bv = b[r0:r1, c0:c1].ravel().astype(np.float64, copy=False)
             bounds = (r0, r1, c0, c1)
 
     finite = np.isfinite(av) & np.isfinite(bv)
